refactor RabbitMQ Serve function to include deferred connection close
